Set enclave initilized bit for target information.
Testing of the report verification code indicated that the
message authentication code being generated for the report was
not consistent with the MAC which was generated in the verifying
enclave.

After significant testing the source of the regression was traced
to the fact that the ->get_targetinfo method was generating
target information which was inconsistent with the state of the
enclave when it was generating the verification key.  In
particular the ->get_targetinfo method was generating enclave
attributes which did not have the initialization bit set.  This
was in turn secondary to the fact the method was obtaining the
attributes from the signature structure rather then the initialized
state of the enclave.

This commit is largely for informational purposes as testing
indicated that the most correct way to handle this situation is
to generate the target information from a null report in the
enclave which is initiating the security context.