Implement enclave report verification.
This update provides an implementation of target enclave report
verification.  This completes the basic infratructure needed to
establish a security context between two enclaves.

The SGX_report structure was modified to separate out the portion
of the report which is validated by the message authenticate code
into a structure named SGX_reportbody.  The report information is
now embedded into the SGX_report structure by an instance of the
SGX_reportbody structure.

A definition was added for the SGX_keyrequest structure which is
used to request enclave specified derived keys.  A definition is
provided for the identifier used to request a report verification
key.

The enclu_egetkey function was added to the SGXfusion library.
This is an assembler based wrapper for calling the ENCLU[EGETKEY]
structure to obtain the various derived keys.

An enclave name LocalSource was created to represent the enclave
which is requesting a secured channel to the LocalTarget enclave.
The sample implementation of report verification was implemented
in this enclave.