Add support for RandomBuffer object.
The primary purpose of this commit was to add the RandomBuffer
object which implements the ability to generate Buffer objects
loaded with random data.

In this initial implementation the source of randomness is the
rdrand processor instruction.  As a result the usual caveats are
available regarding the trustworthiness of the randomness
inherent in this instruction.  Subsequent updates will address
this issue.

The rdrand.s assembler file was added to implement a wrapper
around the call to the processor instruction.  The rdrand()
function implemented in this file returns a boolean true value if
processor randomness was properly returned.

The test harness was updated to provide a framework consistent
with what had been previously implemented in the SGXfusion
library.  The enclave test program is now setup to iterate
through multiple tests.

General additional cleanups were added to the enclave call
interface for the test harness.