Add Curve25519 object.
This update adds access to the Curve25519 object for generating
shared keys using Bernstein's elliptic curve algorithm.

The only modification needed to implement this functionality is
to specify access to the source file in the Makefile.  A rule was
previously implemented that allows source files in the NAAAIM
library to be referenced.

The test harness enclave was updated to encapsulate the test
implemented for the standard userspace implementation as one of
the enclave unit tests.