Implement local enclave key exchange.
This commit implements full support for key exchange between
two local enclaves using the Curve25519 elliptic curve.

The key exchange is implemented by having two separate modes for
the client and the target.  Completion of the two modes leaves
the two enclaves with a shared secret which is secured by the
attestation state of the enclaves.

The source enclave target information is derived from a null
ENCLU[EREPORT] call which means the ->get_targetinfo method is no
longer needed in the enclave object.