Implement query of TPM2 clock information.
This commit provides an initial implementation of support for
returning clock information from the TPM2 device.

The TPM provides two sources of time in milliseconds.  The first
is the current lifespan clock of the device and the second is the
time since the device was last reset.

Counts of the number of device resets and restarts are also
returned.  In addition a field is returned which indicates
whether or not the time is safe, ie. if a value greater then the
current time has been reported.  The value being returned needs
to be verified as correct in the current implementation.

This is not a technically 'safe' implementation of hardware time.
The TPM2 standard refers to a command called GetClock which
provides a signed attestation of time.  Unfortunately the current
Gigabyte platform being tested returns an indication that this
command is not implemented.

This issue will be addressed by attempting to return a signed
quote of a platform configuration register.  In TPM2
implementations all signed attestations carry clock information
which describe the time at which the attestation was generated.