Add support for generating tokens for production enclaves.
The primary focus of this commit is to add support for generating
EINIT tokens for non-debug or production enclaves.  This mode is
selected by the -p command-line arguement to the utility.

This commit also adds a copyright notice to the file and modifies
the usage() function to only print the utility name for the
program banner and copyright lines.