Add support for loading of production enclaves.
This update is focused on providing support for loading
production enclave, ie. enclaves which do not have their debug
bit sit.  This option is selected by the -p command-line
arguement.

In addition to this functionality change a usage function was
added which outputs program header and copyright information, an
optional error message and general usage instructions.