Defend against version 2 metadata.
In the version 2.0 Open Source Gold release it appears that an
incompatible change was made in the SGX metadata.  The size of
the metadata structure was expanded from one to three pages.

In addition the layout group identifiers were changed in a form
which appears to be incompatible with prev-v2.0 enclaves.

This was detected by the change causing a regression in the
SGXmetadata object.  The ->load method was using the section size
to determine the amount of data to copy into the S->metadata
structure.  For an unknown reason the enclave is advertising
version 1.4 metadata compatibility but appears to be linked with
a version 2.0 URTS library which set the section size to three
pages.

This resulted in memory corruption in the ELF loader which caused
a segmentation fault.

A defensive checked was added to generate an error message if
version 2 metadata was detected since these enclaves cannot be
successfully loaded.  In addition the size of the current
metadata structure is used as the amount of data to copy to avoid
the memory overrun.

This combination now allows the pre-built enclaves supplied as
the 'prebuilt-ae-2.0.tar' file to be loaded and initialized.

The ->load method now saves the ELF section size of the metadata
in the state information for the SGXmetadata object.  A METADATA:
section was added to the ->dump method which outputs the actual
section size of the metadata for diagnostic purposes.

The ->load method was also modified to output the metadata
version in MAJOR.MINOR number format to assist in readability.

The sgx-metadata tool was modified to output the enclave name
with an ENCLAVE: header for information purposes.