Initial implementation of provisioning enclave object.
This update implements the initial framework for managing the
Provisioning Enclave.  This is a signed binary enclave provided
by Intel which is used to mediate the provisioning of the EPID
'blob' to an SGX enabled platform.

This blob establishes a trust relationship between Intel and the
platform so that remote verifying parties can use the Intel
Authentication Services (IAS) to verify attestation quotes
generated by the quoting enclave.  In the Intel AESM
implementation EPID blob provisioning occurs the first time a
remote attestation quote is requested for the platform.

Since the SGXfusion architecture is designed in a framework where
individual applications provide all of the SGX infrastructure
support a methodology is needed to conduct the platform EPID
provisioning.  This commit also introduces the sgx-provision
utility which will grow into a command-line utility for
provisioning EPID blobs to a platform at the user's discretion.

The current implementation of the PVE enclave management object
and provisioning tool executes ECALL 3 for the enclave which
implements the creation of the endpoint identifier for the
platform.