Implement message processing object.
This commit implements the SGXmessage object which is being
designed to implement SGX message processing.  This object is
responsible for maintaining a set of type-length-value messages
as a Buffer object.

The ->encode and ->decode methods implement the conversion of a
message header and message list into and out of the ASCII format
used to transact the messages with the Intel application servers.

The PVEenclave object was modified to use this infrastructure to
generate an endpoing selection message.  This message was
confirmed to be able to generate a valid response from the
application servers.

The sgx-provision utility was modified to accept the -1
command-line arguement which takes the response to the endpoint
selection message and dumps the response header and message body.