Updated to sgx-metadata utility.
In order to support submitting documentation to Intel to complete
the SGX licensing process operational updates were implemented
for the sgx-metadata utility.

Option based command-line arguement processing was added.  A mode
based invocation was implemented where the operational mode is
selected with either the -D or -S switches.  The -D mode is the
same as the previous behavior which results in a complete dump of
the SGX metadata.

The -S switch selects signature mode which manages only the
enclave signature data.

The default mode for signature mode is to output the component
members of the SGX_sigstruct structure.  In addition to the
standard output the MRSIGNER value is output which is the SHA256
hash value of the signature module.

Specifiying the -o command-line arguement for signature mode
causes the packed binary representation of the SGX_sigstruct
structure to be output to the designated file.  This is the data
which Intel requests to be submitted for license key whitelisting.