Implement provision response decoding.
The update implements the ->decode method for taking an ASCII
encoded string from the Intel provisiong servers and decomposing
it into a set of Type Length Value (TLV) messages.  Each message
is represented as a single Buffer object.  The list of messages
is managed as an array inside of a new Buffer object which is
included in the state of the object.

The ->dump method was updated to process a response message.  The
set of returned messages is enumerated in the output.

In order to facilitate type checking an enumerated state type was
encoded in the object state.  This has one of three values
indicating whether or not the object is an unitialized state, has
encoded a request or has decoded a response.  Additional checks
need to be added to code to check the status of this variable.