Implement message access methods for SGXmessage object.
This update implements the following new methods for the
SGXmessage object:

	->message_count
	->get_message

The first method is used to determine the number of TLV messages
which were in the server response.  The second method implements
requesting of the return of the payload for a designated message
type.  Requesting a message not found in the current message list
is an object poisoning error.

Since this architecture is designed to support a model where
object clients extract messages that they want to process the
enumerated type which defines all the available messages was
isolated in the following file:

intel-messages.h

This allows clients of the SGXmessage object to include this file
in order to faciliate the request of messages.

The sgx-provision utility was updated to use these new methods to
extract and verify the server endpoint information.  Additional
work needs to be done to implement the appropriate message
verification as only the raw data is now being extracted.