Implement PEK verification.
This update provides the infrastructure to implement verification
of a PEK structure returned by Intel.  The contents of the
structure is verified by a signature using the NIST-ECC256
Digital Signature Algorithm.

The SGXecdsa object was introduced to encapsulate the ECC256
signature functionality.  The goal is to implement an OpenSSL
implemenation of this signature verification by due to how Intel
supplies the public key their infrastructure is used in this
initial implementation.

The sgx-provision tool was updated to use this object to
implement the actual verification of the PEK structure returned
in the first round response message from the Intel provisiong
servers.