Implement endpoint verification.
This update implements verification of the server endpoint
signature.  This is the final step in verifying the endpoint
selection message returned from the Intel servers.

The RSA public key needed to verify the signature supplied in the
endpoint selection message is embedded in the PEK structure which
the previous commit had verified.  The exponent and modulus of
this key are extracted from the verified structure and loaded
into an OpenSSL RSA key object.

The signature block is decrypted and the DER encoding element at
the front of the signature block is verified.  The 32 byte
signature is then extracted from the decrypted signature block.

The signature is verified by comparing it to the following
checksum:

	SHA256(XID || TTL || SERVER_URL)

A successful comparision verifies that communiations is being
conducted with a verified server.