Implement access method for SECS data.
This update is preparatory for work which will be subsequently
committed to support continueing work on EPID provisioning.

That work requires access to parameters of an initialized
enclave.  In order to generically support access to this
information the ->get_secs method was added to the SGXenclave
object.  This method populates a supplied structure with the SECS
data for the enclave.

In order to reflect the initialized state of the enclave the SECS
structure maintained in the state information of the enclave is
updated after successful completion of enclave initialization.
The initialized bit of the attributes data is set and the
mrsigner and mrencave fields are populated.