Add provisioning structures to SGX include file.
This commit adds the following two structures to the generic SGX
definitions:

	SGX_pek

	SGX_extended_epid

The first structure is used to define the RSA key which is used
to authenticate provisioning transactions between the Intel
servers and an SGX endpoint that is being provisioned.

The second structure defines the EPID 'blob' which is the target
of the provisioning exercise.