Add object for AES128-CMAC computations.
This update provides an object to implement the Message
Authentication Code (MAC) that SGX uses for message verification.
It is a block cipher based MAC which uses the AES128 symmetric
cipher.

As was the case with the elliptic curve signature algorithm this
object encapsulates the Intel provided cryptographic primitive
until functionality can be replaced with an OpenSSL based
alternative.