Implement SGX RSA encryption object.
This update implements the SGXrsa object for managing
cryptographic operations needed for EPIC provisioning.  An RSA
key management object is available in the NAAAIM library but is
based on standard PKCSV15 padding using 2048 bit keys.  This
object was implemented to handle the somewhat special
characteristics of the RSA key operations used in EPID
provisioning.

The primary function of the objet is to implement encrypting
using Optimal Assymetric Encryption Padding (OAEP) using a
SHA256 based Feistel generator for the padding.  This is a
decidedly uncommon configuration in current security practices.

The second provisioning message sent to the Intel provisioning
servers uses OAEP SHA256 padding.  Somewhat paradoxically, the
message from the provisioning server which the second message is
generated to in response uses PKCSV15 padding.  So the two
encryption types are not symmetric with respect to their
implementation.

OpenSSL supports this alternative padding but the standard RSA
key object cannot be used as it is restricted to PKCSV15 padding
and OAEP padding using a SHA1 generator.  As a result this object
uses the EVP_PKEY abstraction to implement the encryption
primitives.

As has been previously documented the public key supplied by
Intel is supplied simply as a binary big-endian expression of
both the module and the generator prime.  This object uses the
SGX_pek structure as its initialization arguement as it contains
the public key in this format.