Support creation of the second EPID provisiong message.
This is a rather significant commit which in totality provides
the ability to create the second provisiong message which is sent
to the Intel provisioning servers after endpoint selection is
completed.

This work build on the three cryptographic primities which were
previously commited.

The following methods were added to the PCEenclave in order to
support the extraction of information from the enclaves needed to
support the message:

	->get_ppid

	->get_version

The SGXmessage object was modified to encapsulate all of the
functionality needed to create a single TLV message in the
_encode_message function.  This allows the creation of
sub-messages which can be encapsulated into a single message in
the message stream being returned to the server.

The ->encode_message2 method was added to the SGXmessage object
to drive the creation of the second message.

The ->dump method was modified to print out the description of
TLV messages in a request or response.  The size of the component
messages are printed for debugging purposes as well.

The sgx-provision utility was modified to drive the creation of
the second message using these infrastructure updates.

With these infrastructure updates in place the sgx-provision
utility creates a second provisioning message that is properly
responded to by the Intel provisioning service.