Operability updates to sgx-provision utility.
In order to support multiple message processing various
functional improvements were added to the sgx-provision utility.

Input is now supplied to the utility with the -i switch.

A mode descriptor arguement is now needed to select which type of
message is being generated and/or processed.  The -E switch
indicates that an endpoint selection message is being requested.
The -1 switch specifies processing of the output of the -E
switch and generation of the second message.

The -v switch was added to increase the verbosity level of the
utility.

The -o switch was added to specify a file which the message is to
be output to.  If no output file is specified the message is
directed to the standard output.  Generation of output is now in
the generate_output() function.