Fix input processing.
Converting the sgx-provision utility to using input specified in
files overlooked the need to have the utility open the specified
filenames and read the input.  This was correct by adding the
read_input() function to the utility.

A debug mode was added to the utility with the -D switch.  This
is designed to dump out the contents of a received response for
diagnostic purposes.

Default values were specified for the PVE and PCE tokens since it
made little sense to constantly specify those.  The command line
options were left intact for specifying them as well.