Add initial support for provisioning message 2.
The sgx-provision utility was modified to take the -2
command-line arguement to trigger the processing of provisioning
message 2 that is returned from the provisioning server.

In order to support stage wise debugging the -s switch was added.
This switch is used to specifythe ephemeral key which was used as
the session key for the message that had been used to generate the
message being processed.

The process_message2() function was added to encapsulate high
level processing of the message.  The _decrypt_message2()
function was added as a subordinate function to this function.
This message generates the encryption key that was used by the
provisioning servers to encrypt the internal message carried in
the response.  This key is used along with the server supplied
initialization vector and the protocol header to decrypt and
verify the message.