Implement loading of new message set.
This update implements the ->reload_messages method for the
SGXmessage object.  This method allows the current message list
to be updated with a new set of messages which are supplied as a
binary input buffer.

The motivation for this is to support subordinate messages in a
provisioning message.  The model is to encrypt a block of TLV
messages against the session key generated by the provisioning
client.  This method allows the client to use the decrypted block
of messages to re-populate the message object to allow this
message list to be processed.