Added internal message verification.
The internal message carries a SHA256 hash of the PEK modulus and
generator prime.  This update provides support for computing a
hash of the PEK key that was used to generate the message
predecessor with subsequent comparison of the hash file to the
file in the message.

Support was also added for the -k command-line switch to save or
read a PEK structure.  This provides the ability to save the PEK
from the first provisioning message with subsequent reloading of
the PEK by the function which generates message 2.

The writing and reading of the PEK is implemented in a new helper
function named pek_file().