Add SGX structure size computation utility.
This commit implements the following utility:

sdk-compute-size

Which computes the size of various Intel SGX/SDK data structures
using the Intel header files and include heirarchies.

The architecture of the SGXfusion library is to keep all
structure definitions as statically scoped as possible.  In many
cases such as the creation of provisioning protocol blobs or
enclave communication blobs there is no need to have anything
other then the size of various structures.

This utility reduces the probability of a mistake being made in
computing the size of deeply nested composite structures.