Add message 3 generation method.
This update modifies the object that manages the provisioning
enclave to add the ->get_message3 message which generates the
third message sent to the Intel provisiong servers.

This update adds the SGX_message3 structure to the master SGX
include file.  This structure defines the output message
generated by the provisioning enclave based on input from the
message 2 response.