Remove default alignment for SGX_report.
When used as an address in SGX ENCLU instructions the SGX_report
structure needs to be aligned on a 512 byte boundary.  This
causes when the structure is imbedded in other structures which
must remain compatible with respect to size.

This was noted in the SGX_message3 structure where the SGX_report
structure significantly increased the size of the SGX_message3
structure and created binary compatibility problems when the
structure was passed to enclaves.

The code which was using this to implement local attestation was
modified to add 512 byte alignment directives.