Fix illegal memory reference.
A mistake was made when the read_input() function was added to
the sgx-provision utility.  The strchr() function was used in an
attempt to locate a line delimiter and convert it into a null
character.

This caused an extremely hard to find error which took multiple
hours to locate.  The error would disappear when print statements
were added which made it the quintessential 'Heisenbug'.

On top of this the regression only manifested if a server
response which did not have a line ending character was
processed.  Since there was no null string terminator in the
buffer the search overran the end of the Buffer object and when
it located a value of 0xa that character was changed which resulted
in memory being changed in a non-deterministic fashion.

The regression was fixed by checking for the last character in
the buffer and if that character was a linefeed it was replaced
with a null character.

A safety check was added in the ->decode method of the SGXmessage
object to verify that the decoded binary message was the same
size as the value specified in the response header.