Surface none and key from message two processing.
Creation of the message three response for the Intel provisioning
servers requires access to the nonce and the key that was
generated from that for the incoming message.

The process_message2() function was modified along with the
_decrypt_message2() to surface both of these files upward to code
that is driving the processing of message two and generation of
the message three message.