Use null EPID signature for initial provisioning.
This update sets the EPID signature buffer to NULL with a
corresponding size of 0 in the ->get_message3 method.

The Intel PSW suggests that the EPID signature buffer needs to be
of a minimum size.  Testing has indicated that if a previous
platform information configuration and EPID are being used the
buffer supplied to be populated with an EPID can be NULL with the
signature size set to zero.

A comment in the code provides sizing that will be needed when
processing for a previous EPID is implemented.