Add support for the generation of provisioning message three.
The ->encode_message3 method was added to implement the encoding
of the third provisioning message to the Intel EPID provisioning
service.  This is the last message needed to generate the EPID
blob.

Getting this message encoding operational was problematic until a
regression was noted in the _encode_message private method.  The
size of the encoded message dictates whether a short or long TLV
header is used in a message, with the exception of the TLV_QUOTE_SIG
and TLV_SE_REPORT messages.  For an unknown reason the Intel
defaults these to using a 'long' or four byte header size
encoding.

Without this fix the Intel provisioning servers were returning a
gstatus of GRS_INCORRECT_SYNTAX in response to the improper
encoding of the TLV_SE_REPORT sub-message.