Implement decryption of provisioning message three.
The -3 option was added to the sgx-provision utility to enable
the processing of the third provisioning message returned from
the Intel provisioning servers.  This is the last message in the
chain and contains the EPID group certificate which is the target
of the provisioning exercise.

The desire had been to re-use the message verification and
decryption routines from message 2 but message 3 uses a different
approach to the generation of the Additional Authenticated Data
(AAD).  In addition there is no need to return the nonce so an
alternate function was implemented for clarify of purpose.