Add EPID extraction method.
This commit implements the ->get_epid method which extracts the
EPID blob from the final provisioning message sent by the Intel
provisioning servers.

The input structure for the final ECALL references what Intel
refers to as the psvn_t structure.  This is the first two
elements of the SGX_platform_info structure which is included in
the SGX.h master header file.

A subsequent modification is needed to make the psvn_t structure
a top level construct since it is referenced outside of the
PVEenclave control object as well.