Fix NAAAIM object ordering.
In preparing to add an object to the NAAAIM library I noted there
was an ordering violation in the NAAAIM.h header file.  The
RSAkey and OrgID had the same object numbers which then ascended
upward.

The object library was re-ordered to ensure that all objects were
properly numbered.

There is a need to automate the object numbering since there
should be no dependency on the actual numbers just there
uniqueness.