Implement HTTP object.
This commit provides implementation for an object is used to
manage HTTP transactions with a remote server.  The current
implementation implements only the POST method.

This implementation is currently a wrapper around the wget
utility for the interest of expediency.  The object will be to
implement a native C version as time permits.