Implement direct POST's of protocol messages.
This update uses the HTTP object to implement the direct POSTing
of provisioning messages to the Intel servers.  This is
implemented by placing the code to execute the POST request in
the generate_output() function.  This will automatically cause
all messages which are processed to have their resulting message
forwarded to the provisioning servers for the processing.

Testing of this update also demonstrated that the Intel
provisioning servers 'know' about a host that has previously had
an EPID key provisioned to it.  When message 1 processing is
requested (the output of the endpoing selection message) the
Intel servers will return a provisioning type 3 message which is
the EPID provisioning message for the platform.

This message can be processed with the -3 arguement to the
provisioning tool to yield the  EPID platform key.