Initial implement of quoting enclave object.
This update provides an object named QEenclave that provides an
API for manging the Intel quoting enclave.  This enclave is used
to generate an attestation quote for an enclave running on a
host.  That quote is signed by the EPID private key which has
been provisioned to the platform.

The initial implementation of this object provides an interface
to the ECALL which is used to validate the EPID blob that was
provisioned to the platform.  This is a pre-requisite for using
the enclave to generate a quote.