Add EPID management object.
The changes in this commit introduce an object that will be used
to manage an EPID object.  This is a sealed binary data 'blob'
that is the result of the provisioning process from the Intel
servers.

The blob managed by this object consists of two parts.  A binary
signature created by the Intel group key which includes the private
key provisioned to the host which is specific to the host's
processor.

The second portion is the platform information which is
abstracted from the provisioning process when provisioning
message three is processed.  This is used to determine if the
'blob' is current with the platform or if re-provisioning of the
key which is specific to the processor is needed.