Use SGXepid object to manage EPID.
This commit modifies the PVEenclave object that is used to manage
the provisioning enclave to use the EPID management object to
provision the binary form of the EPID object to the platform.

Before these changes the EPID that was written as a result of the
provisioning process consisted of only the binary data
provisioned from the Intel servers.  The Intel PSW saves the EPID
as a concantenation of that data and the platform information
that is tied to that instance of the EPID.

As part of the message three processing the PVEenclave object now
captures the platform information that is supplied as part of
that message in the EPID management object.  The EPID binary data
is also captured and the EPID management object writes the
concantenated data with the ->save method.