Implement remote attestation test utility.
This commit introduces the initial framework for a utility which
will generate and verify direct remote attestation.  In the model
a client application directly handles the reception and
verification of a quote that is generated by a remote platform.
This process if the preamble for establishing a security context
that can be used to provision secrets to the remote enclave.

The current implementation simply sets up the quoting enclave and
then loads and verifies the EPID that has been provisioned to the
platform.