Update launch token definition.
The previous definition of the launch token structure simply used
a byte sequence to fill out the length of the structure.

One of the data components that will be needed to support
generation of an enclave quote is the platform security
information for an enclave.  This consists of the enclave
security version and the CPU security version.

The launch token structure was updated to include the definition
for all of the fields in the enclave.  This provides an element
definition for the CPU security version that was used to create
the launch token.

The sgx-gen-token utility was updated to use the size of the
structure when writing the output file rather then blithely
assuming it to be 1024 bytes in size.