Track enclave security version.
This update add the SGX_psvn to the set of globally defined data
structures for SGX support.  This data structure encapsulates the
information that defines the trusted computing base of the
enclave.  This information includes the security version of the
enclave and the CPU security version of the initialized enclave.

A byte array structure was added to the enclave state data to
hold the CPU security version.  This information is captured from
the launch token when the enclave is initialized.  The enclave
security version is copied from the enclave metadata after the
enclave is initialized.

The ->get_psvn method was added to the SGXenclave object to
return the enclave security information.