Add platform security version retrieval to PCE enclave.
The ->get_psvn method was added to the object which manages the
PCE enclave.  This is simply a wrapper for a direct call to the
method of the same name in the SGXenclave object.

This lays the framework for verifying whether or not the EPID
blob needs to be re-provision on the platform.