Add support to obtain quoting enclave target information.
The test-remote-target utility was updated to generate target
information for the quoting enclave.  This is another data
element that is needed to support the call to generate the
platform specific enclave quote.

An unused Buffer object variable was removed from the test
utility since it is no longer being referenced.