Add support for generating source report to quoting enclave.
This update adds support for loading the enclave that the
attestation report will be generated for.

The target information for the quoting enclave is provided to the
source enclave which commits and generates a private/public
keypair that will be placed in the report body.  The attestation
process will verify the authenticity of the keypair for the
verifying counter-party.