Add method for generating attestation quote.
This update adds the ->generate_quote method to the QE enclave
object.  This completes the last piece of infrastructure needed
to generate a platform quote against a target enclave.

Support is not fully implemented yet for signature revocation
lists.  The new method will issue an error message if a
revocation list is presented.