Implement generation of remote quote.
The test-remote-target utility was updated in this commit with
the final set of modifications needed in order to generate an
attestation quote.

The first major change was to change from using the LocalSource
enclave to the LocalTarget enclave.  The latter enclave
implements the committment phase of the Curve25519 exchange.

The -s command-line arguement was added in order to specify the
Service Provider ID value that is to be used with the quote.
This is expected to be an ASCII hexadecimal encoding of the 128
byte value assigned by Intel.

Finally the call to the QEenclave->generate_quote message was
added in order to generate the quote.

Currently the utility simply outputs the binary representation of
the quote.