Add quote definition structure.
This update adds the SGX_quote structure which aggregates the
data generated by an enclave attestation quote.  This structure
is also returned by the Intel Attestation Service and the report
data element of the structure will contain, usually, the
Diffie-Hellman key component which is being identified by the
remote attestation.