Fix enclave NONCE and SPID references.
The pointers to the nonce and Service Provider ID that were being
used in the ECALL to generate an attestation quote were
incorrect.  The Intel SDK passes a pointer to a structure which
simply wraps a multi-byte array to the call.  This had caused
some confusion and a pointer to a character array was being used.
This caused the pointer addresses to be used as the nonce and
SPID values rather then the actual byte arrays.

This was fixed by simply passing the pointer from the Buffer
object holding the values into the ECALL.