Implement full remote attestation processing.
This update makes significant changes to the test-remote-target
utility in order to implement full processing of an enclave
attestation quote with generation of an attestation report from
the Intel Attestation Services.

The attestation report is encoded in JSON format.  The extraction
of specific fields is implemented with regular expressions so
there is a generic mechanism for processing tagged entries.

Processing of the attestation report also includes interpretation
of the Platform Information Blob which is returned if the EPID
group has been revoked or is out of date.  This documents the
exact nature of the attestation services interpretation of the
trusted computing base of the platform.