Convert SHA256 object to SGX/OpenSSL.
This update introduces general support for the SGX version of the
OpenSSL library whose port was done by Intel.  The SHA256 object
was modified to use the OpenSSL SHA256 support rather then the
Intel cryptography library.

The SHA256 object is functionally identical to the NAAAIM library
but a modified version was needed to address the API change
between OpenSSL pre-1.1 and the 1.1.x release that the Intel port
is based on.

The version of SGX/OpenSSL which this work is being done against
is a heavily modified variant of the Intel code.  In particular
support for threading was dropped in order to simplify the
interface with untrusted userspace.  The current OpenSSL version
is segmentation faulting when the OpenSSL_add_all_digests()
function is called or any other function which leads to generic
initialization of the OpenSSL library.